requirements.txt
playwright>=1.40.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pandas==2.2.1
tqdm>=4.66.0  # for progress bars
openai>=1.0.0
flask==3.0.2
flask-cors==4.0.0 